move interface to a seperate file
