+++ src/app/app.component.ts
@@ -1,21 +1,6 @@
 import { Component } from '@angular/core';
-import desseretData from '../../public/data.json';
 import { AddToCartComponent } from "./components/add-to-cart/add-to-cart.component";
 
-// interface
-interface Dessert {
-  image: DessertImages;
-  name: string;
-  category: string;
-  price: number;
-};
-
-interface DessertImages {
-  thumbnail: string;
-  mobile: string;
-  tablet: string;
-  desktop: string;
-};
 
 @Component({
   selector: 'app-root',
@@ -27,9 +12,9 @@ interface DessertImages {
 
 export class AppComponent {
   title = 'Product list';
-  desserts:Dessert[] | null = null;
+  // desserts:Dessert[] | null = null;
 
-  constructor() {
-    this.desserts = desseretData;
-  };
+  // constructor() {
+  //   this.desserts = desseretData;
+  // };
 };
