+++ src/app/models/dessert.interface.ts
@@ -1,0 +1,14 @@
+// interface
+interface Dessert {
+  image: DessertImages;
+  name: string;
+  category: string;
+  price: number;
+};
+
+interface DessertImages {
+  thumbnail: string;
+  mobile: string;
+  tablet: string;
+  desktop: string;
+};
